include new method in location
