+++ Data/Products/ProductRepository.cs
@@ -218,23 +218,6 @@ namespace Viam.SalesForce.API.Data.Products
             }
         }
 
-        public ActionResult<List<AgencyProducts>> getAgencyProducts(string idProduct, string date, string idLocation)
-        {
-            using (IDbConnection dbConnection = Connection)
-            {
-                dbConnection.Open();
 
-                var p = new DynamicParameters();
-
-                p.Add("@IdProduct", idProduct);
-                p.Add("@IdLocation", idLocation);
-				p.Add("@Date", date);
-
-                var query = dbConnection.Query<AgencyProducts>(Helper.Constants.spGetSLDailyTransactions,
-                                                          p,
-                                                          commandType: CommandType.StoredProcedure);
-                return query.AsList();
-            }
-        }
     }
 }
